backend/requirements.txt
django>=5.2,<5.3
djangorestframework>=3.16,<3.17
django-cors-headers>=4.7,<4.8
python-dotenv>=1.1,<1.2
dj-database-url>=3.0,<3.1
psycopg[binary]>=3.2,<3.3
djangorestframework-simplejwt>=5.3,<5.4
